speech-processor/requirements.txt
kafka-python
pymongo
requests
python-socketio
numpy
librosa